add AmazonStability3Config
